+++ ui/mods/com.pa.quitch.gwaioverhaul/gw_play/card_units.js
@@ -1400,5 +1400,8 @@ define({
     {
       id: "gwaio_upgrade_subcommander_duplication",
     },
+    {
+      id: "gwaio_upgrade_subcommander_fabber",
+    },
   ],
 });