+++ ui/mods/com.pa.quitch.gwaioverhaul/gw_play/cards.js
@@ -303,6 +303,7 @@ if (!gwaioCardsLoaded) {
               "gwaio_upgrade_storm",
               "gwaio_upgrade_stryker",
               "gwaio_upgrade_subcommander_duplication",
+              "gwaio_upgrade_subcommander_fabber",
               "gwaio_upgrade_sxx",
               "gwaio_upgrade_teleporter",
               "gwaio_upgrade_torpedolauncher",