+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - New tech cards:
   - Sub Commander Duplication Tech
+  - Sub Commander Fabber Tech
 
 ## v5.24.2 - 2021-11-19